Integer to Long

+++ src/main/java/ch/heigvd/pclist/services/dao/CpuDAO.java
@@ -82,7 +82,7 @@ public class CpuDAO implements CpuDAOLocal {
         return cpuList;
     }
 
-    public List<Cpu> get(int pageSize, int pageIndex) {
+    public List<Cpu> get(long pageSize, long pageIndex) {
         List<Cpu> cpuList = new ArrayList<>();
 
         try {
@@ -91,8 +91,8 @@ public class CpuDAO implements CpuDAOLocal {
                     "FROM cpu " +
                     "LIMIT ? OFFSET ?;");
 
-            preparedStatement.setInt(1, pageSize);
-            preparedStatement.setInt(2, pageSize * pageIndex);
+            preparedStatement.setLong(1, pageSize);
+            preparedStatement.setLong(2, pageSize * pageIndex);
             ResultSet resultSet = preparedStatement.executeQuery();
 
             while (resultSet.next()) {
